requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-celery-results==2.6.0

celery==5.5.3
redis==7.0.1

openai==2.8.1
requests==2.32.5
python-dotenv==1.2.1

langchain==1.0.7
langchain-core==1.0.5
langchain-openai==1.0.3
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43

pydantic==2.12.4
pydantic-core==2.41.5
orjson==3.11.4
ormsgpack==1.12.0
tiktoken==0.12.0

python-dateutil==2.9.0.post0
sniffio==1.3.1
anyio==4.11.0
httpx==0.28.1
httpcore==1.0.9
typing-extensions==4.15.0
packaging==25.0
six==1.17.0
tenacity==9.1.2
